Styling: reorder code
Move code out of dunder-init.
