+++ gaphor/core/styling/tests/test_selector.py
@@ -1,6 +1,6 @@
 import pytest
 
-from gaphor.core.styling import parse_style_sheet
+from gaphor.core.styling import compile_style_sheet
 
 
 class Node:
@@ -57,7 +57,7 @@ def test_node_test_object_child_parent():
 def test_select_all():
     css = "* {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector("any")
 
@@ -65,7 +65,7 @@ def test_select_all():
 def test_select_name():
     css = "classitem {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(Node("classitem"))
     assert not selector(Node("packageitem"))
@@ -74,7 +74,7 @@ def test_select_name():
 def test_select_inside_combinator():
     css = "classitem nested {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(Node("nested", parent=Node("classitem")))
     assert selector(
@@ -90,7 +90,7 @@ def test_select_inside_combinator():
 def test_select_parent_combinator():
     css = "classitem > nested {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(Node("nested", parent=Node("classitem")))
     assert not selector(
@@ -106,7 +106,7 @@ def test_select_parent_combinator():
 def test_attributes():
     css = "classitem[subject] {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(Node("classitem", attributes={"subject": "val"}))
     assert not selector(Node("classitem"))
@@ -116,7 +116,7 @@ def test_attributes():
 def test_attribute_equal():
     css = "classitem[subject=foo] {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(Node("classitem", attributes={"subject": "foo"}))
     assert not selector(Node("classitem", attributes={"subject": "bar"}))
@@ -125,7 +125,7 @@ def test_attribute_equal():
 def test_attribute_in_list():
     css = "classitem[subject~=foo] {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(Node("classitem", attributes={"subject": "foo"}))
     assert selector(Node("classitem", attributes={"subject": "foo bar"}))
@@ -135,7 +135,7 @@ def test_attribute_in_list():
 def test_attribute_starts_with():
     css = "classitem[subject^=foo] {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(Node("classitem", attributes={"subject": "foo"}))
     assert selector(Node("classitem", attributes={"subject": "foomania"}))
@@ -145,7 +145,7 @@ def test_attribute_starts_with():
 def test_attribute_starts_with_dash():
     css = "classitem[subject|=foo] {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(Node("classitem", attributes={"subject": "foo"}))
     assert selector(Node("classitem", attributes={"subject": "foo-mania"}))
@@ -155,7 +155,7 @@ def test_attribute_starts_with_dash():
 def test_attribute_ends_with():
     css = "classitem[subject$=foo] {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(Node("classitem", attributes={"subject": "foo"}))
     assert selector(Node("classitem", attributes={"subject": "manicfoo"}))
@@ -165,7 +165,7 @@ def test_attribute_ends_with():
 def test_attribute_contains():
     css = "classitem[subject*=foo] {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(Node("classitem", attributes={"subject": "foo"}))
     assert selector(Node("classitem", attributes={"subject": "be foo-ish"}))
@@ -175,7 +175,7 @@ def test_attribute_contains():
 def test_attributes_with_dots():
     css = "classitem[subject.members] {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(Node("classitem", attributes={"subject.members": "foo"}))
 
@@ -183,7 +183,7 @@ def test_attributes_with_dots():
 def test_attributes_are_lower_case():
     css = "ClassItem[MixedCase] {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(Node("classitem", attributes={"mixedcase": "foo"}))
 
@@ -191,7 +191,7 @@ def test_attributes_are_lower_case():
 def test_empty_pseudo_selector():
     css = ":empty {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(Node("node"))
     assert specificity == (0, 1, 0)
@@ -201,7 +201,7 @@ def test_empty_pseudo_selector():
 def test_empty_pseudo_selector_with_name():
     css = "node:empty {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(Node("node"))
     assert not selector(Node("node", children=[Node("child")]))
@@ -215,7 +215,7 @@ def test_empty_pseudo_selector_with_name():
 def test_hovered_pseudo_selector(state):
     css = f":{state} {{}}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(Node("node", state=(state,)))
     assert selector(Node("node", state=(state, "other-state")))
@@ -225,7 +225,7 @@ def test_hovered_pseudo_selector(state):
 def test_is_pseudo_selector():
     css = "classitem:is(:hover, :active) {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(Node("classitem", state=("hover",)))
     assert selector(Node("classitem", state=("active",)))
@@ -237,7 +237,7 @@ def test_is_pseudo_selector():
 def test_not_pseudo_selector():
     css = "classitem:not(:hover) {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(Node("classitem"))
     assert selector(Node("classitem", state=("active")))
@@ -249,7 +249,7 @@ def test_not_pseudo_selector():
 def test_has_pseudo_selector():
     css = "classitem:has(nested) {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(Node("classitem", children=[Node("nested")]))
     assert specificity == (0, 0, 2)
@@ -262,7 +262,7 @@ def test_has_pseudo_selector():
 def test_has_pseudo_selector_with_complex_selector():
     css = "classitem:has(middle > nested) {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(
         Node("classitem", children=[Node("middle", children=[Node("nested")])])
@@ -285,7 +285,7 @@ def test_has_pseudo_selector_with_combinator_is_not_supported():
     # able to deal with this. This test is just here to illustrate.
     css = "classitem:has(> nested) {}"
 
-    error, payload = next(parse_style_sheet(css))
+    error, payload = next(compile_style_sheet(css))
 
     assert error == "error"
 
@@ -293,7 +293,7 @@ def test_has_pseudo_selector_with_combinator_is_not_supported():
 def test_has_and_is_selector():
     css = "node:has(:is(:hover)) {}"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector(
         Node(
@@ -306,7 +306,7 @@ def test_has_and_is_selector():
 def test_media_query():
     css = "@media(prop = true) { diagram { color: blue; } }"
 
-    (selector, specificity), payload = next(parse_style_sheet(css))
+    (selector, specificity), payload = next(compile_style_sheet(css))
 
     assert selector
     assert specificity == (0, 0, 1)
